backend/requirements.txt
celery>=5.5.3
redis>=4.5.1
mysql-connector-python>=8.0.0
python-dotenv>=0.21.1
requests>=2.31.0